chore: use hidden method
